ios: refactor RNUnityProxy
The old approach results in linker error when building for Device SDK,
since an implementation of RNUnity can't be found at link time, which
is true. The approach works when targetting Simulator SDK, and used to
work on Xcode 12 and older versions I believe.

The new approach adds a method to pass a pointer to a type capable of
emitting React Native events from outside UnityFramework, which means
we no longer need to rely on linker. Our Unity plugin has been updated
to use this pointer when emitting events from Unity.