feature: Keep Awake
There are other React Native modules to use this feature[1], but
unfortunately we get race condition on Android because two pieces of
asynchronous code are manipulating window flags at the same time, and it
might work or not depending on the device you're using.

This fixes the problem by implementing Keep Awake directly in this
module so that we can add some extra glue on Android and make sure
things will work as intended.

* Use the keepAwake prop on UnityView to turn this on and off

1: e.g. @sayem314/react-native-keep-awake