android: Another race condition with keep awake
Keep awake is sometimes turned off after a few seconds. Check this every
second for a couple of seconds and turn it on again if necessary.

The problem was observed both in the example app and a separate app
using proprietary Unity content, and this patch fixes it in both apps.